Display delete confirmation modal before deleting a comment and simulate the deletion of attraction. Implement NgbActiveModal within DeleteConfirmationComponent. Enhance confirmation by also incorporating window.confirm before deletion.

+++ wclient/src/app/comment/comment-list/comment-item/comment-item.component.ts
@@ -2,6 +2,8 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Attraction } from '../../../attraction/attraction.model';
 import { Comment } from '../../comment.model';
 import { CommentService } from '../../comment.service';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { DeleteConfirmationComponent } from '../../../shared/delete-confirmation/delete-confirmation.component';
 
 @Component({
   selector: 'app-comment-item',
@@ -13,9 +15,26 @@ import { CommentService } from '../../comment.service';
 export class CommentItemComponent {
   @Input() comment!: Comment;
 
-  constructor(private commentService: CommentService) {}
+  constructor(
+    private commentService: CommentService,
+    private modalService: NgbModal
+  ) {}
 
   onDeleteComment(commentId: number) {
-    this.commentService.deleteComment(commentId);
+    const modalRef = this.modalService.open(DeleteConfirmationComponent);
+    modalRef.componentInstance.selectedReason = '';
+    modalRef.componentInstance.otherReason = '';
+
+    console.log(modalRef.result);
+
+    modalRef.result
+      .then((result) => {
+        if (result === 'delete') {
+          this.commentService.deleteComment(commentId);
+        }
+      })
+      .catch((reason) => {
+        console.log('Modal dismissed due to: ', reason);
+      });
   }
 }
